+++ src/frontend/js/components/Form/Field.stories.tsx
@@ -1,0 +1,16 @@
+export const defaultConfig = {
+  args: {
+    label: 'Any label here',
+    message: 'Any message here',
+    fieldClasses: [],
+    id: 'id',
+  },
+  argTypes: {
+    error: {
+      control: 'boolean',
+    },
+    message: {
+      control: 'text',
+    },
+  },
+};