+++ src/frontend/js/components/Icon/index.stories.tsx
@@ -1,0 +1,11 @@
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { Icon } from './index';
+
+export default {
+  title: 'Components/Icon',
+  component: Icon,
+} as ComponentMeta<typeof Icon>;
+
+const Template: ComponentStory<typeof Icon> = (args) => <Icon {...args} />;
+
+export const Default = Template.bind({});